Adding new script to repo
This script reads every pdf file in the directory it is run in, converts them all to plaintext,
and outputs those files with a name to match the original that it read from.

Useful when you have a ton of pdf files that you want to search, and you have some idea what words
you want to search for, but you don't want to waste time searching every single one.

You can use this, search all files (CTRL+SHIFT+F) then find which files contain the text you're looking for.